fix: ()の定義をoperatorから分割するように
