+++ src/tokenize/state/in_normal/by_parenthesis.c
@@ -1,0 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   by_parenthesis.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: surayama <surayama@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2026/01/17 00:00:00 by surayama          #+#    #+#             */
+/*   Updated: 2026/01/17 00:00:00 by surayama         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "in_normal_internal.h"
+
+void	by_parenthesis(t_token_store *store, t_token_state *state,
+		char current)
+{
+	if (push_token(store) == ERROR || add_buffer(store, current) == ERROR
+		|| push_token(store) == ERROR)
+		*state = ON_ERROR;
+	else
+		*state = IN_NORMAL;
+}
