+++ src/tokenize/state/in_normal/in_normal_internal.h
@@ -1,0 +1,25 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   in_normal_internal.h                               :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: surayama <surayama@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2026/01/17 00:00:00 by surayama          #+#    #+#             */
+/*   Updated: 2026/01/17 00:00:00 by surayama         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef IN_NORMAL_INTERNAL_H
+# define IN_NORMAL_INTERNAL_H
+
+# include "tokenize.h"
+
+void	by_last(t_token_store *store, t_token_state *state, char current);
+void	by_space(t_token_store *store, t_token_state *state, char current);
+void	by_quote(t_token_store *store, t_token_state *state, char current);
+void	by_operator(t_token_store *store, t_token_state *state, char current);
+void	by_parenthesis(t_token_store *store, t_token_state *state,
+			char current);
+
+#endif
